refactor: 경기 결과 도메인 패키지 명 match로 변경
