+++ src/main/java/kr/tennispark/match/admin/infrastructure/MatchResultRepository.java
@@ -1,6 +1,6 @@
-package kr.tennispark.record.admin.infrastructure;
+package kr.tennispark.match.admin.infrastructure;
 
-import kr.tennispark.record.common.domain.entity.MatchResult;
+import kr.tennispark.match.common.domain.entity.MatchResult;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MatchResultRepository extends JpaRepository<MatchResult, Long> {
