+++ src/main/java/kr/tennispark/match/admin/infrastructure/MemberRecordRepository.java
@@ -1,6 +1,6 @@
-package kr.tennispark.record.admin.infrastructure;
+package kr.tennispark.match.admin.infrastructure;
 
-import kr.tennispark.record.common.domain.entity.association.MatchParticipation;
+import kr.tennispark.match.common.domain.entity.association.MatchParticipation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRecordRepository extends JpaRepository<MatchParticipation, Long> {
